+++ app/Http/Controllers/Billing/GasInvoiceController.php
@@ -173,7 +173,7 @@ class GasInvoiceController extends Controller
             'payable_amount' => 0,
             'paid_amount' => 0,
             'balance_amount' => 0,
-            'due_date' => Carbon::now()->addDays(Constants::DPNG_DUEDAYS->value)->format('Y-m-d'),
+            'due_date' => Carbon::now()->addDays((int)Constants::DPNG_DUEDAYS->value)->format('Y-m-d'),
             'status_id' => InvoiceStatus::PAID->value, // paid
             'created_by' => Auth::id()
         ];
@@ -243,7 +243,7 @@ class GasInvoiceController extends Controller
             'payable_amount' => $inv_total,
             'paid_amount' => NULL,
             'balance_amount' => $inv_total,
-            'due_date' => Carbon::now()->addDays(Constants::DPNG_DUEDAYS->value)->format('Y-m-d'),
+            'due_date' => Carbon::now()->addDays((int)Constants::DPNG_DUEDAYS->value)->format('Y-m-d'),
             'status_id' => InvoiceStatus::NOT_PAID->value, // Not paid
             'created_by' => Auth::id()
         ];
@@ -326,7 +326,7 @@ class GasInvoiceController extends Controller
             'payable_amount' => $inv_total,
             'paid_amount' => NULL,
             'balance_amount' => $inv_total,
-            'due_date' => Carbon::now()->addDays(Constants::DPNG_DUEDAYS->value)->format('Y-m-d'),
+            'due_date' => Carbon::now()->addDays((int)Constants::DPNG_DUEDAYS->value)->format('Y-m-d'),
             'status_id' => InvoiceStatus::NOT_PAID->value, // Not paid
             'created_by' => Auth::id()
         ];
